[SandboxIRVec] Fix canonicalization of LLVM ShuffleVectorInst for UnpackInst
Vector UnpackInst's used to have a poison value as their operand(0).
But when generating an UnpackInst's LLVM IR, we would place the poison
at operand 1.
Having the poison at operand 1 makes the masks more readable so this is the
canonicalization we are going to keep.

Test based on test provide by @slackito.