[SandboxVec][Scheduler] ReadyList comparator
This patch implements the hierarchical comparator for the ready list.
Since the dependency graph does not create dependency edges from PHIs or
to terminator nodes to save memory and compile-time, we need to force
these dependencies in the scheduler.

It's the job of the ready list to prioritize instructions such that
PHIs always come first and terminators always come last.